Rename spot_skills Python package to spot_skills_py
