+++ src/spot_skills/launch/pose_estimation_demo.launch
@@ -27,4 +27,8 @@
     <!-- Launch RViz with a demo-specific config file -->
     <node name="rviz" pkg="rviz" type="rviz" args="-d $(find spot_skills)/rviz/pose_estimate.rviz"/>
 
+    <!-- Launch rqt_gui to visualize the images used for pose estimation -->
+    <node name="rqt_gui" pkg="rqt_gui" type="rqt_gui"
+        args="--perspective-file $(find spot_skills)/rviz/pose_estimate.perspective"/>
+
 </launch>
