+++ src/spot_skills/setup.py
@@ -6,7 +6,7 @@ from catkin_pkg.python_setup import generate_distutils_setup
 
 # Fetch values from package.xml
 setup_args = generate_distutils_setup(
-    packages=["spot_skills", "transform_utils"],
+    packages=["spot_skills_py", "transform_utils"],
     package_dir={"": "src"},
 )
 setup(**setup_args)
