+++ src/spot_skills/src/spot_skills_py/planning_scene/put_down_surface.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from spot_skills.tamp.geometry.real_range import RealRange
+    from spot_skills_py.tamp.geometry.real_range import RealRange
 
 
 @dataclass
