+++ src/spot_skills/src/spot_skills_py/tamp/geometry/real_range.py
