+++ src/spot_skills/src/spot_skills_py/tamp/samplers/grasp_pose_sampler.py
@@ -2,8 +2,8 @@
 
 import numpy as np
 
-from spot_skills.tamp.geometry.real_range import RealRange
-from spot_skills.tamp.geometry.transforms_3d import Point3D, Pose3D, Quaternion
+from spot_skills_py.tamp.geometry.real_range import RealRange
+from spot_skills_py.tamp.geometry.transforms_3d import Point3D, Pose3D, Quaternion
 
 from .sampler import Iterator, Sampler
